video-detector/requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
attrdict==2.0.1
colorama==0.4.3
easydict==1.9
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
flake8-import-order==0.18.1
grpcio==1.51.1
importlib-metadata==1.6.0
jdcal==1.4.1
joblib==0.14.1
lap==0.4.0
mccabe==0.6.1
more-itertools==8.2.0
motmetrics==1.2.0
openpyxl==3.0.3
packaging==20.3
pandas==1.0.3
Pillow==8.3.2
pluggy==0.13.1
protobuf==4.21.12
py==1.10.0
py-cpuinfo==5.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
pytest==5.4.1
pytest-benchmark==3.2.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.4
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
torch==1.4.0
torchvision==0.5.0
Vizer==0.1.5
wcwidth==0.1.9
xmltodict==0.12.0
zipp==3.1.0